+++ resources/js/components/Register.jsx
@@ -2,10 +2,14 @@
 import React, { useState } from "react";
 import axios from "axios";
 import Swal from "sweetalert2";
+import { auth } from "../firebase";
+import { RecaptchaVerifier, signInWithPhoneNumber } from "firebase/auth";
 import { useNavigate, Link } from "react-router-dom";
 
 const Register = () => {
     const navigate = useNavigate();
+    const [verificationCode, setVerificationCode] = useState("");
+    const [confirmationResult, setConfirmationResult] = useState(null);
     const [formdata, setFormData] = useState({
         name: "",
         email: "",
@@ -21,44 +25,47 @@ const Register = () => {
         e.preventDefault();
         try {
             const response = await axios.post("/api/register", formdata);
-            //   const phoneNumber = `+94${response.data.user.phone}`;
+            const phoneNumber = `+94${response.data.user.phone}`;
 
-            //   const recaptchaVerifier = new RecaptchaVerifier(
-            //     auth,
-            //     "recaptcha-container",
-            //     { size: "normal" }
-            //   );
+            const recaptchaVerifier = new RecaptchaVerifier(
+                auth,
+                "recaptcha-container",
+                { size: "invisible" }
+            );
 
-            //   const confirmationResult = await signInWithPhoneNumber(
-            //     auth,
-            //     phoneNumber,
-            //     recaptchaVerifier
-            //   );
-            //   setConfirmationResult(confirmationResult);
+            const confirmationResult = await signInWithPhoneNumber(
+                auth,
+                phoneNumber,
+                recaptchaVerifier
+            );
+            setConfirmationResult(confirmationResult);
 
             // Handle user input for OTP
-            //   const code = prompt("Enter OTP sent to your phone");
-            //   setVerificationCode(code);
+            const code = prompt("Enter OTP sent to your phone");
+            setVerificationCode(code);
 
-            //   const result = await confirmationResult.confirm(code);
-            //   const idToken = await result.user.getIdToken();
+            const result = await confirmationResult.confirm(code);
+            const idToken = await result.user.getIdToken();
 
-            //   const verifyResponse = await axios.post(`/api/2fa`, { idToken });
-            //   if (verifyResponse.data.message === "2FA verified") {
-            const token = response.data.token;
-            localStorage.setItem("token", token);
-            // localStorage.setItem("user", JSON.stringify(response.data.user));
-            Swal.fire({
-                icon: "success",
-                title: "Login Successful",
-                text: "Welcome back!",
-            }).then(() => {
-                navigate("/dashboard");
-            });
-            //     navigate("/dashboard");
-            //   } else {
-            //     console.error("2FA verification failed");
-            //   }
+            const verifyResponse = await axios.post(`/api/2fa`, { idToken });
+            if (verifyResponse.data.message === "2FA verified") {
+                const token = response.data.token;
+                localStorage.setItem("token", token);
+                // localStorage.setItem("user", JSON.stringify(response.data.user));
+                Swal.fire({
+                    icon: "success",
+                    title: "Login Successful",
+                    text: "Welcome back!",
+                }).then(() => {
+                    navigate("/dashboard");
+                });
+            } else {
+                Swal.fire({
+                    icon: "error",
+                    title: "Login Failed",
+                    text: "2FA verification failed",
+                });
+            }
         } catch (error) {
             if (error.response && error.response.status === 401) {
                 Swal.fire({
@@ -153,8 +160,11 @@ const Register = () => {
                                     </form>
                                     <p className="text-center text-muted mt-5 mb-0">
                                         Have already an account?{" "}
-                                        <Link to="/login" className="fw-bold text-body">
-                                        Login here
+                                        <Link
+                                            to="/login"
+                                            className="fw-bold text-body"
+                                        >
+                                            Login here
                                         </Link>
                                     </p>
                                 </div>
